Removed incorrectly added methods to user
