+++ app/models/user.rb
@@ -22,31 +22,10 @@ class User < ApplicationRecord
   validates :password_confirmation, presence: true, on: :create
 
   before_save :to_lowercase
-  after_commit :update_is_bloggable_status
 
   private
 
     def to_lowercase
       email.downcase!
-    end
-
-    def upvotes_count
-      votes.upvote.count
-    end
-
-    def downvotes_count
-      votes.downvote.count
-    end
-
-    def net_votes
-      upvotes_count - downvotes_count
-    end
-
-    def update_is_bloggable_status
-      threshold = 10
-      new_status = net_votes >= threshold
-      if is_bloggable != new_status
-        update_column(:is_bloggable, new_status)
-      end
     end
 end
